add option to skip file if it exists

+++ cryptodataset/max.py
@@ -59,12 +59,15 @@ class MAXData(Base):
         df['datetime'] = pd.to_datetime(df['timestamp'], unit='ms')
         return df
 
-    def download_ohlcv(self, symbol: str, timeframe: str, output_dir: Path) -> None:
-        df = self.get_ohlcv(symbol, timeframe)
-
+    def download_ohlcv(self, symbol: str, timeframe: str, output_dir: Path, skip: bool = False) -> None:
         output_dir.mkdir(parents=True, exist_ok=True)
         csv_path = output_dir / 'MAX_{}_{}.csv'.format(symbol.replace('/', '').upper(), timeframe)
 
+        if skip and csv_path.exists():
+            logger.info('{} already exists, skip', csv_path)
+            return
+        
+        df = self.get_ohlcv(symbol, timeframe)
         logger.info('saving ohlcv to {}', csv_path)
         df.to_csv(csv_path, index=False)
